@@ -5,11 +5,23 @@
   "keywords": [],
   "main": "src/index.js",
   "dependencies": {
+    "@emotion/react": "^11.4.1",
+    "@emotion/styled": "^11.3.0",
     "@material-ui/core": "4.12.3",
     "@material-ui/icons": "4.11.2",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
-    "react-scripts": "3.2.0"
+    "@mui/icons-material": "^5.0.3",
+    "@mui/material": "^5.0.3",
+    "bootstrap": "^5.1.1",
+    "bootstrap-icons": "^1.5.0",
+    "dotenv": "^10.0.0",
+    "firebase": "^9.1.1",
+    "react": "^17.0.0",
+    "react-bootstrap": "^1.6.4",
+    "react-dom": "^17.0.0",
+    "react-router-dom": "^5.3.0",
+    "react-scripts": "^4.0.3",
+    "supervillains": "^3.0.0",
+    "unique-names-generator": "^4.6.0"
   },
   "devDependencies": {
     "typescript": "3.3.3"
