@@ -2,11 +2,14 @@ import React, { useState } from "react";
 import AddIcon from "@material-ui/icons/Add";
 import Fab from "@material-ui/core/Fab";
 import Zoom from "@material-ui/core/Zoom";
+import { useAuth } from "../contexts/AuthContext";
 
 function CreateArea(props) {
+  const { currentUser } = useAuth();
   const [note, setNote] = useState({
     title: "",
-    content: ""
+    content: "",
+    userId: currentUser.uid
   });
 
   function handleChange(event) {
@@ -21,13 +24,17 @@ function CreateArea(props) {
   }
 
   function submitNote(event) {
+    
     props.onAdd(note);
     setNote({
       title: "",
-      content: ""
+      content: "",
+      userId: currentUser.uid
     });
     event.preventDefault();
   }
+
+  // Note edit effects
   const [isTyping, setIsTyping] = useState(false);
 
   return (
@@ -46,7 +53,7 @@ function CreateArea(props) {
           onChange={handleChange}
           onClick={() => setIsTyping(true)}
           value={note.content}
-          placeholder="Take a note..."
+          placeholder={!isTyping? `Welcome ${currentUser.displayName||"Mr. Anonymous"}! Take a note here...`:"Take a note..."}
           rows={isTyping ? 3 : 1}
         />
         <Zoom in={isTyping}>
