@@ -2,4 +2,11 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "./components/App";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+
+
+ReactDOM.render(
+  
+    <App />
+  ,
+ document.getElementById("root")
+ );
